requirements.txt
Flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Babel==4.0.0
Babel==2.13.1
httpx==0.25.2
stripe==7.8.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
